Adjusted types and moved files from rebase on master.
